backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
python-dotenv==0.19.0
pydantic==1.8.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
openai==0.27.0
websockets==10.1
alembic==1.7.7
livekit-server-sdk==0.3.0
fastapi-swagger==0.1.1
